requirements.txt
# FastAPI
fastapi
uvicorn
python-multipart
python-dotenv

# LangChain
langchain
langchain-community
openai
pypdf
tiktoken
faiss-cpu

# CORS
aiofiles  
httpx    
